Move `FoldersViewModel` to `:legacy:ui:folder`
