+++ legacy/ui/folder/build.gradle.kts
@@ -8,5 +8,12 @@ android {
 
 dependencies {
     implementation(projects.core.ui.legacy.designsystem)
+
+    implementation(projects.legacy.account)
     implementation(projects.legacy.folder)
+    implementation(projects.legacy.mailstore)
+    implementation(projects.legacy.message)
+    implementation(projects.legacy.search)
+
+    implementation(libs.androidx.lifecycle.livedata.ktx)
 }
