+++ legacy/ui/folder/src/main/java/app/k9mail/legacy/ui/folder/FoldersViewModel.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.ui.folders
+package app.k9mail.legacy.ui.folder
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
@@ -8,8 +8,6 @@ import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.mailstore.FolderRepository
 import app.k9mail.legacy.message.controller.MessageCountsProvider
 import app.k9mail.legacy.search.SearchAccount
-import app.k9mail.legacy.ui.folder.DisplayUnifiedInbox
-import app.k9mail.legacy.ui.folder.FolderList
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
